Got all the transfer COLP/CONP information working properly now.
There is one little restriction...if the transferer is attempting
to transfer the transferee to an unbridged call (e.g. transfering
a person to an IVR), then we have to indicate the connected line
update on the channel directly from the SIP monitor thread, meaning
that we cannot run a connected line macro on the channel.