Get app_rpt compiling again.  I doubt seriously that it actually works.
Also, the code in this module is horrendous and we should remove it from the
tree.  I'm not sure who is supposed to be maintaning this thing, but they
clearly are not.  I don't see the sense of leaving it in the main tree.  If it
lives *anywhere* it should be in addons.