Fixes several call transfer issues with chan_misdn.
JIRA ABE-1903 - Orphaned held calls left in music-on-hold.

Do not simply block passing the hangup event on held calls to asterisk
core.

issue #14692 - Held calls are not automatically cleared after transfer.

Allow the core to initate disconnect of held calls to the ISDN port.  This
also fixes a similar case where the party on hold hangs up before being
transferred or taken off hold.

issue #14355 - Crash if attempt to transfer a call to an application.

Masquerade the other pair of the four asterisk channels involved in the
two calls.  The held call already must be a bridged call (not an
applicaton) or it would have been rejected.

Fixed to allow held calls to be transferred to ringing calls.  Previously,
held calls could only be transferred to connected calls.

*  Eliminated unused call states to simplify hangup code.
*  Eliminated most uses of "holded" because it is not a word.