A branch for an RTP streaming backend for res_monitor (1.4)
A new branch to add channel monitoring to a remote server as RTP
streams. The recordings are intended to be sent to an Oreka/Orex server.

Metadata about the RTP stream is sent in dummy SIP INVITE and BYE
messages.

This branch includes the code developed vs. 1.4 as this is the code that
is actually tested. A branch based on trunk will be available soon.