format_rtp: a "file format" to dump an RTP stream
Add a new format module to dump frames into a simple "context-less" RTP
stream.

Unlike the existing RTP APIs, we don't want to assume that the RTP
stream is part of some VoIP session, and would like to avoid any setup.
The remote party here is completely passive and will not respond to a
single packet we send.