Provide means for monitors to register with the core.
This provides a way for a channel driver to register
a device monitor type so that the appropriate type
of monitor may be created when CC is requested. I need
to modify app_dial so that the frame data of an
AST_CONTROL_CC frame contains the type of monitor
that should be created.

A similar mechanism will be needed for agents, too.
I glossed over this step earlier since I am focusing
strictly on generic agents for now, though. With monitors
I can't do that though. It may be possible to combine
the registration of agents and monitors, although I don't
think I will do that just in case for some reason
someone should wish to only make a monitor or only
make an agent for a specific channel driver. Eh,
I'll think about it.