Get CC configuration parameters onto the monitor.
This is necessary so that decisions can be made regarding
the cc_max_monitors option, and the cc_available_timer can
be armed easily.

In making these changes, I've noticed a couple of things to
improve on:

1. ast_cc_interface and ast_cc_monitor have a lot of overlap.
I am going to isolate the common elements into a single struct
for ease-of-use and ease-of-reading.

2. There is a lot of allocation and copying of ast_cc_config_params.
It may be a better idea to make use of reference-counting instead.

I can hopefully knock both of these out in the next couple of commits.