Fix some potential refcount bugs.
There is no way to cancel a taskprocessor task, and so there
is no way of preventing a taskprocessor task which has a pointer
to a freed monitor from running. The fix is to bump the refcount
of the monitor up before pushing the task, and then unreffing
when the task has completed.