Part 1 of handling max_cc_monitors.
The next part will be to actually increment and decrement
the num_requests portion of an ast_cc_monitor as needed.
This should be fairly easy, since the number of requests
will correspond to the number of parent links a monitor has.

Create a link, increase num_requests.
Destroy a link, decrease num_requests.