A workaround to send the final BYE even though we have no CDR
At hangup, chan_cdr is released in ast_hangup() and the monitoring is
stopped a bit later in the destructor of the channel object.

Thus when we get to monitor2rtp_stop(), we no longer have chan->cdr .

Previously I added a sanity check to just bail out in such a case. Now I
replace it with providing some dummy CDR data. Hopefully there will be a
better solution.