Change how CC is offered to a caller.
Now, CC is offered based on the cc_agent_policy of
the caller. The dialplan writer is capable of changing
this setting prior to calling the Dial app, so it is
still possible to administratively disable CC from
being used. This makes things a bit more clear and
from a user perspective, makes things "just work"