This branch is dedicated to issue #15911 and all the good times it shall bring.
ast_do_masquerade does not follow proper locking order in trunk. Since the name
of a channel may change during the masquerade, and the name is used as the hash
value for the channel's ao2 container, the channel has to be unlinked and then
re-linked with the new hash value.  Unlinking/linking lockes the container while
the channel is already locked... This is not proper locking order. If a channel
lock is held while attempting to obtain the container lock, deadlock madness will
ensue.