This is pretty sweet.
Okay, so part of the CC work involved is the ability to configure CC
parameters in channel device configuration sections. Not all channel
drivers have had the necessary code updates to parse these config options.
This change will make it so that even if the channel driver has not been
updated to parse the necessary configuration options, the configuration
parameters can still be set on a channel using the CALLCOMPLETION
dialplan function.

The reason why is that the CALLCOMPLETION function searches for the
config parameters datastore on the channel. The only reason it would
not be present is if the channel's driver had not been updated to
parse the CC config settings. So what we do now is, if we don't find
the datastore, we just create it with the defaults. That way it can
be retrieved and the configuration items can be set properly on it.