Add flags to CC agents.
There was a check in the code which would specifically
see if the agent being used was generic. Well, while that
made sense, what makes more sense is to check whether a
specific flag is set instead. That way, other agent types
who might also be affected can also set that flag and things
will just work. It also makes the code more clear too.

There's still one other place where the check being used
is to see if the agent policy is set to generic, but I'm
leaving that one because at the time, the agent has not
yet been created.