Add an offer_cc callback for ast_cc_agents.
With ISDN, the calling channel driver needs to be made
aware of the availability of CC as well as the interfaces
that may accept CC prior to sending back the ALERTING
message in the case of CCNR or prior to sending back the
DISCONNECT in the case of CCBS.

Prior to the addition of this callback, the calling channel
driver was not made aware of the availability of CC until
after all information from each callee had been collected.

The tricky part with this change is that the agent itself
is not allocated until a core instance has been allocated.
Agent creation time will need to occur earlier, perhaps when
the first AST_CONTROL_CC frame is read during a call. I'll
probably simplify things and create the core instance there
as well.