Branch for solving this issue involve app_queue and chan_local: ---------------------------------------------------------------
We have a Queue() that calls a local/xx@yy that calls a SIP channel with dial(SIP/annika,,t)

The local channel masquerades away with the "t" and/or "T", so the bridge option disappears and the
callee or caller can no longer transfer after the masquerade.

In this branch, I'm experimenting with sending a new control frame, AST_CONTROL_BRIDGEPARAM
upstream, so that the queue bridge gets an update and updates the bridge parameters for the
callee, based on the original parameters in the dying poor old deprecated bridge downstream.

------
Co-developed with the Danish national team "Chan_local rocks!"