Fix a bug introduced when the core allocation was changed to at the reception of the first CC control frame.
The bug was that the AST_FLAG_OFFER_CC flag was never set on
the inbound channel, meaning that the offer timer was never
started.

This has been corrected by adding a flag to the dialed_cc_interfaces
datastore which indicates if CC is offerable to the channel. The logic
works such that only the channel on which the datastore was originally
created will have this flag set true. Channels which inherit the datastore
will have the flag cleared.