Branch to attempt to fix the SIP forking support that we've lost in chan_sip.
During SIPit we discovered that the SIP channel has lost support of forked calls
with the integration of the "spiral" patch. As this is required in most of
the large installations with SIP proxys involved, we need to fix this.

kpfleming has been working on a long term solution for 1.6 or trunk.

I will use the fallback RFC-2543 compatible mode described in RFC 3261:

"The INVITE request matches a transaction if the Request-URI , To tag, From tag, Call-ID, CSeq, and top
Via header field match those of the INVITE request which created the transaction."
(section 17.2.3).

Basically this means that if the Requested URI (extension) is not the same in two
INVITE messages, it's not the same call, even if we have the same Call-ID and
to/from tags etc. kpfleming is working on more RFC 3261-compliant support,
separating on the topmost, leftmost via branch tag.

I hope that this patch will be smaller and easier to swallow for fixing 1.4 and maybe
be used as a short-time fix for the other branches.

This work is purely founded by Edvina AB.