Initial work towards supporting PUBLISH in chan_sip.
For now, I've put structures in place for defining a general
purpose event state compositor, as well as some fundamental
code for handling incoming PUBLISH requests.

The code is intended to work as a base so that if we wish to
support PUBLISH for event types other than call-completion,
it is obvious how it should be done, and it shouldn't be
too difficult to do so either.