Simplify the event_publication_agent a bit.
After thinking through the use-cases a bit, there's not a good reason
to maintain a container of EPA entries within an EPA.

When an application needs to transmit a SIP PUBLISH, it can maintain a
reference to the EPA entry that it has created. When a response to a
PUBLISH comes back, the sip_pvt will have a reference to the EPA entry.

In other words, whenever something needs to refer to an EPA entry, it will
or at least should have the reference handy. It's not like the ESC, which will
need to look up an entry based on information in an incoming request.