Fix some incorrect hashing in SIP ESCs.
ESC entries are hashed based on the entity tag that
is created for them. There were two problems fixed here.

1. The ESC entries were being linked into the ESC before
having an entity tag assigned. The result would be that
we would never be able to find an ESC entry since we would
look in the incorrect bucket.

2. Since we create a new entity tag for every 200 OK we
send, the result is that an ESC entry's entity tag would
change from what it was when originally linked. To fix this,
we have to unlink, change the entity tag, and then re-link.