Fix some behavior for the SIP CC agent destructor.
It's possible to have the function called before we have
received a SUBSCRIBE. In such a case, we need to be sure that
the subscribe_pvt is non-NULL before trying to act on it.

Also, if failure occurs before we have sent a response to the
SUBSCRIBE, we need to send a 500 back.