Stuff.
Now we will not register the SIP CC agent if XML support
is not available. In addition, if a peer is defined with a "native"
cc_agent_policy, it will be changed to "never" and a warning will
be emitted explaining why. Also, al CC ESC code will not be compiled
if XML support is non-existent.

A user may still set a SIP caller's agent policy to "native" using
the CALLCOMPLETION dialplan function. However, when it comes time
to attempt creation of the agent during a call, the CC engine will
be unable to since the SIP CC agent was not registered. The result
will be that the CC transaction will be immediately torn down. In other
words, things will work fine.