Generate silence in between CNG tones.
Previously, the loop that generates the CNG tone sent no other audio in between
tones.  This resulted in a failure to detect CNG when the receiver was some
some versions of Asterisk.  The cause of this failure is that the CNG detection
code will not report successful detection of the tone until it sees the tone
end.  Since no other audio was being sent, the detection was never reported.

The code now enables a silence generator on the channel while CNG is not being
sent.  CNG is now properly detected on the receiving side.