Implement a new agent callback, stop_ringing.
The comment above the callback in ccss.h explains its purpose
and use pretty well. It's basically one of those odd edge cases
that only will be needed when Asterisk is used as the TE side
of an ISDN PTMP call, and likely won't even be needed then. However,
the possibility exists and so it is important that we provide the
necessary means to do what needs to be done.