Prevent consecutive duplicates from appearing in CC_INTERFACES.
If dialing, say, a group of DAHDI channels, it's possible that
there are multiple devices or PRI spans that will be monitored
to determine when a member of that group has become available.
When the time comes, we don't want to end up with a CC_INTERFACES
variable with contents like

DAHDI/G1&DAHDI/G1&DAHDI/G1

Instead, just one will suffice.