Add a new parameter to initreqprep to provide an explicit request URI.
When sending CC-related NOTIFY and PUBLISH requests, we need to set the
request URI to a specific value that the other end is expecting, rather than
any other settings on the sip_pvt. This new change allows for such a URI
to be passed in.

For now, all callers of initreqprep are setting this parameter NULL. My next
set of changes should include their usage of the parameter.