Branch for named ACL's for trunk.
A named ACL is an Access Control List that can be manipulated after configuration and live in it's own name space. The NACL module manage a list of NACL objects that can be used by other modules, like channel drivers, manager and dialplan apps.

Several SIP devices can share the same access control list and there will be one
for the whole SIP channel. An external application that reads the security events
can manipulate the NACLs in real time through AMI and block/unblock devices.

There's also an API so that Asterisk modules can modify NACLs internally.

Among the future ideas are NACLs that can be set by referring to a DNS name and use the DNSmgr to stay up to date with DNS. That requires some changes to the ACL.c api that will happen in the trunk version only.
I have also been playing with the idea of having a callback so that an app will know when a NACL is matched or some sort of counters to measure activity per time period and trigger alarms. Kamailio has one implementation of something like this in the pike module. 

Anyway, a lot of ideas has been based on named ACLs, so I thought that was a starting point and a good holiday hack :-)

Feedback and comments is, as always, welcome.

/O