Add more logic into get_destination in chan_sip.
This makes it so that a CC recall will have the recall core ID returned
as an output parameter to handle_request_invite. After a channel gets
created for the INVITE, we can then call the proper functions to set up
datastores and set the CC_INTERFACES channel variable.

In addition, this commit also adds a missing '*' to the prototype for
ast_cc_agent_callback in ccss.h