Creating a sub-branch of CCSS to restructure how monitors work.
I'm doing this in a sub-branch of CCSS because for the most part CCSS works
how it is, plus I don't want to mess with internal things too much while Richard
is still developing ISDN stuff.

The reason for this branch is that the monitor setup in CCSS is just too plain
complicated, and furthermore there is a core issue that cannot be fixed especially
easily. I'll enumerate some of the problems that Richard and I have come across
while doing the work.

* The idea of having a single monitor structure per device (as opposed to one per
device per call) works wonderfully for generic monitoring, but it makes very little
sense for SIP and ISDN monitoring. The result is that the monitor structure just
isn't used in those channel drivers and they have to create their own container
of monitor instances instead.

* There is a lot of information duplicated in the interface list that is stored
on the agent and in the monitors and links.

* Extension monitors are really not that useful. Their callbacks are always just
used to pass information down to device monitors or up to their parent extension
monitors.

* There is some somewhat evil code needed where we call the monitor instance_destructor
callback in places where an outbound channel driver will allocate CC-related data
but a monitor has not yet been allocated. This will be minimized with this restructuring.

* There is a bit of an ambiguity if multiple destinations are called by a single channel
driver. Both monitors created will have the same core ID, and so when an instance destructor
is called, it is not possible to determine which monitor instance is to be destroyed. This
was sort of the straw that broke the camel's back, because unlike other problems encountered,
there's no good way to work around this one.

I have decent plan of attack in place, and so I hope that this work doesn't take much longer
than a week to complete.