Next batch of progress that compiles.
First, I've changed the interface list to be doubly-linked. The
main reason for this is to be able to do the types of traversals
I'll need when calling monitor callbacks. For an example of
how it is used, see the request_cc function in ccss.c

I'll be using the template that I've made in request_cc for
the other instances where I call monitor callbacks next. For
now, the new monitor callbacks are getting a "2" appended to
their names. Once I remove the original monitor callbacks,
the "2" will be removed.