A few changes to aid generic monitoring.
* Got rid of a circular reference in a generic_monitor_instance.

* Changed ast_cc_monitor_callee_available2 to ast_cc_monitor_callee_available
and changed its parameter from a monitor to a core_id. This helps the generic
monitor and other implementations have no need for a potential circular reference.

* Fixed some faulty logic regarding suspension and unsuspension of monitors. The
is_suspended flag is not actually used by the core, and so it makes more sense
for that to exist on the generic_monitor_instance structure instead. I also now
make sure to set it as appropriate and actually clear it when the monitor becomes
unsuspended.