Change when we call cancel_available_timer.
The problem I had was when I tried to test canceling CC. The problem
was that the scheduler still held a reference to my monitor from when
the available timer was started. Due to when cancel_available_timer
was called, the result always a crash. Now we call cancel_available_timer
before unreffing the monitor either from the interface list destructor
or from cc_monitor_failed.