Add code to deal with monitor suspension in the case of multiple call legs.
Most of the code previously committed makes the assumption
that when monitoring a SIP endpoint, it is the only one
that was called. However, this is not always the case, and
so smarts have to be built in to know how to handle specific
situations.

This commit deals with the biggest of the offenders, the suspension
and unsuspension routines. There is still a bit more to do here,
I suspect.