Improve code having to do with SIP call-completion.
This is mostly just some cleanup. Returning more accurate
values, handling failure cases properly, etc. There is
also a grievous logic error fixed in the monitor's suspend
callback.

I'm currently having an issue where my PUBLISH dialogs are not
being destroyed, apparently because they are not being unlinked
from the container of dialogs. The problem I'm having is that
I am unclear on what the proper chan_sip way of doing this is.
What's also odd is that I don't see this for my SUBSCRIBE requests,
which are handled much the same way. Alas, this is a problem
for tomorrow.