Fix logic with regards to ringing and CC in app_dial.
The large XXX comment in the source sums it up well, but
here's the short version. Basically, we can't send a ringing
frame to the caller unless we have either received a CC frame
from one of the called parties, all called parties have sent
ringing frames, or CC is disabled for the call.